requirements.txt
torch
torchvision
transformers
scikit-learn
matplotlib
numpy
pandas
pillow
tqdm
seaborn
safetensors
